rename index.js -> parse.js
